Porting application to C#
After writing the launcher in C#, I decided that it would be better for
me as an amateur programmer to program in a different language than the
one I am familiar (Java). From now on, commits will be in Visual Studio
sln files. This solution is compiled (for now) with .Net Framework 4.
When the time comes, I will compile with Mono 2/3 for cross-platform
compatibility.